ecommerce_extensions/payment/pp_sandbox/requirements.txt
certifi==2020.12.5
chardet==3.0.4
Django==2.2
djangorestframework==3.12.2
idna==2.10
pytz==2020.4
requests==2.25.0
sqlparse==0.4.1
urllib3==1.26.2
zappa==0.52.0
